+++ database/migrations/tenant/2025_12_22_084327_m_farmer.php
@@ -20,8 +20,8 @@ return new class extends Migration
             $table->string('ph_no',15)->nullable();
             $table->string('wp_no',15)->nullable();
             $table->enum('open_type', ['give', 'get'])->nullable();
-            $table->unsignedBigInteger('open_bal')->nullable();
-            $table->unsignedBigInteger('adv_prime')->nullable();
+            $table->integer('open_bal')->nullable();
+            $table->integer('adv_prime')->nullable();
             $table->string('acc_type',30)->nullable();
             $table->string('b_name',100)->nullable();
             $table->string('acc_name',100)->nullable();
@@ -29,7 +29,7 @@ return new class extends Migration
             $table->string('ifsc',30)->nullable();
             $table->string('upi',50)->nullable();
             $table->unsignedInteger('fav')->default(0); // if count > 0 then favorite
-            $table->unsignedBigInteger('c_by')->nullable(); 
+            $table->integer('c_by')->nullable(); 
             $table->enum('status', ['active', 'inactive'])->default('active');
             $table->timestamps();
         });
