+++ database/migrations/tenant/2025_12_22_090219_loss_category.php
@@ -15,7 +15,7 @@ return new class extends Migration
             $table->id();
             $table->string('loss',200)->nullable();
             $table->enum('status', ['active', 'inactive'])->default('active');
-            $table->unsignedBigInteger('c_by')->nullable(); 
+            $table->integer('c_by')->nullable(); 
             $table->timestamps();
         });
     }
