+++ database/migrations/tenant/2026_01_11_181139_create_e_invoice_table.php
@@ -13,15 +13,15 @@ return new class extends Migration
     {
         Schema::create('e_invoice', function (Blueprint $table) {
             $table->id();
-            $table->unsignedBigInteger('inv_id')->nullable();
-            $table->unsignedBigInteger('load_id')->nullable();
-            $table->unsignedBigInteger('product')->nullable();
+            $table->integer('inv_id')->nullable();
+            $table->integer('load_id')->nullable();
+            $table->integer('product')->nullable();
             $table->decimal('total', 10, 2)->nullable()->default(0);
             $table->decimal('grace',10, 2)->nullable()->default(0);
             $table->decimal('price',10, 2)->nullable()->default(0);
             $table->decimal('bill_amt',10, 2)->nullable()->default(0);
             $table->enum('status', ['active', 'inactive'])->default('active');
-            $table->unsignedBigInteger('c_by')->nullable();
+            $table->integer('c_by')->nullable();
             $table->timestamps();
         });
     }
