+++ database/migrations/tenant/2026_01_11_181759_create_e_party_table.php
@@ -13,12 +13,12 @@ return new class extends Migration
     {
         Schema::create('e_party', function (Blueprint $table) {
             $table->id();
-            $table->unsignedBigInteger('party_id')->nullable();
+            $table->integer('party_id')->nullable();
             $table->enum('type', ['pay_in', 'pay_out'])->default('pay_in');
             $table->decimal('amount', 10, 2)->default(0);
             $table->string('method', 100)->nullable();
             $table->enum('status', ['active', 'inactive'])->default('active');
-            $table->unsignedBigInteger('c_by')->nullable();
+            $table->integer('c_by')->nullable();
             $table->timestamps();
         });
     }
