+++ database/migrations/tenant/2026_01_11_184115_create_petty_cash_table.php
@@ -13,13 +13,13 @@ return new class extends Migration
     {
         Schema::create('petty_cash', function (Blueprint $table) {
             $table->id();
-            $table->unsignedBigInteger('emp_id')->nullable();
+            $table->integer('emp_id')->nullable();
             $table->enum('type', ['petty', 'settle'])->default('petty');
             $table->decimal('amount', 10, 2)->nullable();   
             $table->string('method', 250)->nullable();
             $table->date('date')->nullable();
             $table->enum('status', ['active', 'inactive'])->default('active');
-            $table->unsignedBigInteger('c_by')->nullable();
+            $table->integer('c_by')->nullable();
             $table->timestamps();
         });
     }
